* Improving support for adding custom.css so that the file is also managed properly during the plugin update process * Updating localization files
git-svn-id: https://plugins.svn.wordpress.org/comment-images/trunk@638957 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ plugin.php
@@ -4,7 +4,7 @@ Plugin Name: Comment Images
 Donate URI: http://tommcfarlin.com/donate/
 Plugin URI: http://tommcfarlin.com/comment-images/
 Description: Allow your readers easily to attach an image to their comment.
-Version: 1.6
+Version: 1.6.1
 Author: Tom McFarlin
 Author URI: http://tommcfarlin.com/
 Author Email: tom@tommcfarlin.com
@@ -44,8 +44,13 @@ class Comment_Image {
 		// Load plugin textdomain
 		add_action( 'init', array( $this, 'plugin_textdomain' ) );
 		
-		// Setup the activation hook specifically for checking for the custom.css file
+		/* Setup the activation hook specifically for checking for the custom.css file
+		 * I'm calling the same function using the activation hook - which is when the user activates the plugin,
+		 * and during upgrade plugin event. This ensures that the custom.css file can also be managed
+		 * when the plugin is updated.
+		 */
 		register_activation_hook( __FILE__, array( $this, 'activate' ) );
+		add_action( 'pre_set_site_transient_update_plugins', array( $this, 'activate' ) );
 	
 		// Determine if the hosting environment can save files.
 		if( $this->can_save_files() ) {
@@ -85,7 +90,7 @@ class Comment_Image {
 		 $str_custom_path =  dirname( __FILE__ ) . '/css/custom.css';
 		 
 		 // If the custom.css file doesn't exist, then we create it
-		 if( ! file_exists( $str_custom_path ) ) {
+		 if( is_writable ( $str_custom_path ) && ! file_exists( $str_custom_path ) ) {
 			 file_put_contents( $str_custom_path, '' );
 		 } // end if
